Implement connection handle management in WebSocket connections. Add a handle field to the Connection struct and a set_handle method for setting the connection handle during setup. Update relevant areas in the connection lifecycle to utilize the handle, ensuring consistent access throughout. Enhance documentation for clarity on connection management.
